create detect package
